refactor(code_gen): extract load_module into modules.rs
